Extend the about section and split api file into folder
