+++ client/public/index.html
@@ -8,7 +8,7 @@
     <meta name="theme-color" content="#000000" />
     <meta
       name="description"
-      content="A tool for calculating and comparing unit stats for Warhammer Age of Sigmar"
+      content="A tool for calculating and comparing damage outputs for Warhammer Age of Sigmar units (also referred to as Mathhammer)."
     />
     <link rel="apple-touch-icon" href="logo192.png" />
     <!--
