+++ client/public/static/about.md
@@ -1,7 +1,26 @@
 # AoS Statshammer
 
 A tool for calculating and comparing damage outputs for
-Warhammer Age of Sigmar units.
+Warhammer Age of Sigmar units (also referred to as Mathhammer). Some of the features include:
+
+- Average Damage Stats
+- Profile Modifiers (Abilities)
+  - Examples:
+    - *'Deal `x` Mortal Wounds on a roll of `y`'*
+    - *'Reroll rolls to `x`'*
+    - etc.
+- Target Modifiers
+  - Examples:
+    - *'Ethereal'*
+    - *'Negate wounds and mortal wounds on a roll of `y`'*
+- Simulations
+    - Generating `x` simulations to gather:
+      - Discrete Probability
+      - Cumulative Probability
+      - Sample Metrics: mean, max, variance, etc.
+    - The number of simulations are customizable
+- Import/Export JSON
+- And more
 
 This is a free community tool built by Damon Hook (`NoMaDhOoK`)
