+++ client/src/api/index.ts
@@ -1,0 +1,3 @@
+export * from './compare';
+export * from './modifiers';
+export * from './simulations';
