+++ client/src/components/Footer/Footer.tsx
@@ -51,7 +51,8 @@ const Footer: React.FC = () => {
     <footer className={classes.footer}>
       <Paper className={clsx(classes.paper)}>
         <Typography variant="body2" component="p">
-          Built by: Damon Hook
+          Built by: Damon Hook&nbsp;
+          <i>(NoMaDhOoK)</i>
         </Typography>
         <Typography variant="body2" component="p">
           Disclaimer: This tool is in no way endorsed or sanctioned by Games Workshop - it is unofficial and
